Renaming AlgorithmFactory to HMACSHAAlgorithmFactory
